requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.7
APScheduler==3.10.4
pydantic==2.5.2
pydantic-settings==2.1.0
loguru==0.7.2
httpx==0.25.2
python-dotenv==1.0.0 